refactor: move test helpers to common folder
